keep track of the next entity id
